created elasticService which will handle all calls to elastic search
