+++ plugins/gui/controllers/dashboardController.js
@@ -1,12 +1,13 @@
 var jwt = require('jsonwebtoken');
 var logger = require('../utils/logger');
 var mailer = require('../mail/mailer');
+var elasticService = require('../services/elasticService.js');
 
 module.exports.show = function (req, res) {
     //Get the users email, name, and profile picture
     var domain = req.user.split("@")[1];
     if (domain == 'sofi.org' || domain == 'sofi.com') {
-        res.render('dashboard');
+        res.render('dashboard', {currentUrl: ''});
     } else {
         req.logout();
         res.status(401);
@@ -14,7 +15,23 @@ module.exports.show = function (req, res) {
     }
 };
 module.exports.showAbout = function(req, res) {
-    res.render('about');
+    res.render('about', {currentUrl: ''});
+};
+module.exports.showRetries = function(req, res) {
+    elasticService.poll(req.body.service, function(err, logs) {
+        if (logs == null) {
+            req.flash('error', err);
+        }
+        res.render('retries', {currentUrl: 'retries', logs: logs})
+    });
+};
+module.exports.showReplays = function(req, res) {
+    elasticService.poll(req.body.service, function(err, logs) {
+        if (logs == null) {
+            req.flash('error', err);
+        }
+        res.render('replays', {currentUrl: 'replays', logs: logs})
+    });
 };
 module.exports.sendDailySummary = function (req, res) {
     var authHeader = req.headers['authorization'];
