+++ plugins/gui/controllers/elasticSearchController.js
@@ -7,44 +7,6 @@ var client = new elasticsearch.Client({
     host: (process.env.FK_ES_HOST || 'localhost') + ":" + (process.env.FK_ES_PORT || 9200)
 });
 
-module.exports.poll = function(req, res) {
-    var service = req.body.service;
-    var logs = [];
-    var index = 'forklift-'+service+'*';
-
-    client.search({
-        index: index,
-        size: 100,
-        body: {
-            query: {
-                query_string: {
-                    query: "Error",
-                    fields: ["step"]
-                }
-            },
-            "sort": [{
-                "time": {"order": "desc"}
-            }]
-        }
-    }).then(function (resp) {
-        resp.hits.hits.forEach(function(hit) {
-            logs.push(hit);
-        });
-        var json = {
-            log: true,
-            response: logs
-        }
-        res.end(JSON.stringify(json));
-    }, function(err) {
-        logger.error(err.message);
-        var json = {
-            log: false,
-            response: err
-        }
-        res.end(JSON.stringify(json));
-
-    });
-};
 module.exports.updateAsFixed = function(req, res) {
     var updateId = req.body.id;
     var index = req.body.index;
