+++ plugins/gui/package.json
@@ -13,6 +13,7 @@
     "dateformat": "^1.0.12",
     "elasticsearch": "^10.1.2",
     "express": "~4.12.4",
+    "express-session": "^1.14.2",
     "forever-monitor": "^1.7.0",
     "fs": "0.0.2",
     "jade": "^1.11.0",
