+++ plugins/gui/public/stylesheets/401.css
@@ -11,7 +11,7 @@ p {
 }
 
 h3 {
-    color: whitesmoke;
+    color: #000000;
     font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
     font-size: 3em;
     font-weight: lighter;
