+++ plugins/gui/views/404.jade
@@ -2,7 +2,7 @@ html
     head
         meta(name='viewport', content='width=device-width, initial-scale=1.0')
         link(href='css/bootstrap.min.css', rel='stylesheet', media='screen')
-        link(href='css/console.css', rel='stylesheet', media='screen')
+        link(href='css/401.css', rel='stylesheet', media='screen')
         link(href='css/font-awesome.min.css', rel='stylesheet', media='screen')
 
         <!-- Navbar -->
@@ -23,7 +23,9 @@ html
         div(class="dashboard-container")
             div(id="console")
                 div(id="status")
-                    h3 404 <p class="red">Page not found!</p>
+                    h3 404
+                        br
+                        .red.text-uppercase this page does not exist
 
     script(src='js/bootstrap.min.js')
