+++ plugins/gui/views/login.jade
@@ -18,5 +18,5 @@ html
                             <span class="fa fa-google"></span> Sign in with Google
                             </a>
 
+    script(src='js/bootstrap.min.js')
     script(src='js/jquery-1.11.3.min.js')
-    script(src='js/docs.js')
